+++ compiler-core/src/format/tests/bit_array.rs
@@ -202,3 +202,13 @@ fn block_value() {
 "#
     );
 }
+
+#[test]
+fn operator_in_pattern_size() {
+    assert_format!(
+        "pub fn main() {
+  let assert <<len, payload:size({ len + 1 } * 8 + 1)>> = <<>>
+}
+"
+    );
+}