+++ compiler-core/src/javascript/expression.rs
@@ -1593,7 +1593,7 @@ impl<'module, 'a> Generator<'module, 'a> {
         Ok(docvec![left, " ", op, " ", right])
     }
 
-    fn bin_op_with_doc_operands(
+    pub(super) fn bin_op_with_doc_operands(
         &mut self,
         name: BinOp,
         left: Document<'a>,
@@ -1619,11 +1619,11 @@ impl<'module, 'a> Generator<'module, 'a> {
                 docvec!["remainderInt", wrap_args([left, right])]
             }
             BinOp::DivInt => {
-                self.tracker.int_remainder_used = true;
+                self.tracker.int_division_used = true;
                 docvec!["divideInt", wrap_args([left, right])]
             }
             BinOp::DivFloat => {
-                self.tracker.int_remainder_used = true;
+                self.tracker.float_division_used = true;
                 docvec!["divideFloat", wrap_args([left, right])]
             }
         }